requirements.txt
python-telegram-bot==12.4.2 
requests==2.22.0
urllib3==1.25.7
pycricbuzz==2.4
